pyproject.toml
[project]
name = "pyasiaph-2026"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "django-browser-reload>=1.21.0",
    "django-compressor>=4.5.1",
    "django-extensions>=4.1",
    "django-watchfiles>=1.4.0",
    "psycopg2-binary>=2.9.10",
    "pydantic-settings>=2.11.0",
    "sentry-sdk>=2.39.0",
    "wagtail>=7.1.1",
    "whitenoise>=6.11.0",
]

[dependency-groups]
dev = [
    "pre-commit>=4.3.0",
    "ruff>=0.13.2",
]
